Replace local diff crate with similar
