+++ Cargo.toml
@@ -24,7 +24,6 @@ edition = "2021"
 [dependencies]
 tokio = { version = "1", features = ["net", "time", "macros", "io-util", "rt"] }
 rust_spell = { path = "rust_spell" }
-diff = { path = "rust/diff" }
 rust_buffer = { path = "rust_buffer" }
 rust_python = { path = "rust_python" }
 rust_profiler = { path = "rust_profiler" }
@@ -41,6 +40,7 @@ rust_blob = { path = "rust_blob" }
 rust_sha256 = { path = "rust_sha256" }
 rust_blowfish = { path = "rust_blowfish" }
 regex = "1"
+similar = "2"
 
 [features]
 default = []
