--- rust/diff/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "diff"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-similar = "2"
-[lib]
-name = "diff"
-path = "src/lib.rs"
-crate-type = ["cdylib", "rlib"]
